+++ relay-server/src/processing/spans/process.rs
@@ -247,7 +247,10 @@ fn normalize_span(
         if ctx.is_processing() {
             eap::normalize_ai(&mut span.attributes, duration, model_metdata);
         }
-        relay_event_normalization::normalize_performance_score(span, performance_score);
+        // In the old pipeline, the profile would get saved in the performance score context.
+        // In the new pipeline, we are not storing it for now.
+        let _ = relay_event_normalization::normalize_performance_score(span, performance_score);
+        eap::normalize_mobile_measurements(&mut span.attributes, duration);
         eap::normalize_attribute_values(&mut span.attributes, allowed_hosts);
         eap::write_legacy_attributes(&mut span.attributes);
         eap::normalize_client_sample_rate(&mut span.attributes);